Move more stuff into shared file
